+++ social-network-ui/src/components/Chats/Chats.jsx
@@ -1,18 +1,21 @@
 import React, { useState, useEffect } from "react";
 
 import { useDispatch, useSelector } from "react-redux";
-import { getMembers, setChatId } from "../../redux/slices/chatSlice";
+// import { getMembers, setChatId } from "../../redux/slices/chatSlice";
 import UseUserToken from "../../hooks/useUserToken";
 import ChatMembers from "../ChatMembers/ChatMembers";
 import {
-  fetchChats,
-  getChatMessages,
+  // fetchChats,
+   setChatId,
   setMessages,
-  createMessage,
-  setError,
+  // createMessage,
+  // setError,
 } from "../../redux/slices/chatSlice";
+import {getChatMessages} from '../../api/getChatMessages';
+
 import { Box } from "@mui/material";
 import PropTypes from "prop-types";
+import {fetchChats} from '../../redux/thunks/chatThunk'
 
 function Chats() {
   const { token } = UseUserToken();
@@ -20,13 +23,16 @@ function Chats() {
   //  const [chats, setChats] = useState(null)
   const dispatch = useDispatch();
   const chats = useSelector((state) => state.chat.chats);
-
+  //  const userId = useSelector(state => state.user?.userId)
+  //  console.log(userId);
+console.log(chats);
   const error = useSelector((state) => state.chat.error);
   useEffect(() => {
     dispatch(fetchChats(token));
   }, [dispatch, token]);
 
   async function fetchMessages(chatId) {
+    console.log(chatId);
     const data = await getChatMessages(chatId, token);
 
     dispatch(setMessages(data));
@@ -44,9 +50,11 @@ function Chats() {
     >
       {/* <div onClick={() => {}} style={{width: '100%'}}> */}
       {chats?.map((chat) => (
+   
         <ChatMembers
+         key={chat.id}
           fetchMessages={fetchMessages}
-          key={chat.id}
+         
           chatmembers={chat}
         />
       ))}
