+++ social-network-ui/src/redux/slices/chatSlice.js
@@ -3,11 +3,15 @@ import { createSlice } from "@reduxjs/toolkit";
 import UseUserToken from "../../hooks/useUserToken";
 import getUserId from "../../utils/getUserId";
 import axios from "axios";
+import {getChats} from '../../api/getChats'
+import {getChatMessages} from '../../api/getChatMessages'
+import {getChatMembers} from '../../api/getChatMembers'
+import {fetchChats} from '../thunks/chatThunk'
 
 const userId = getUserId()
 
 const initialState = {
-  chats: null,
+  chats: [],
   error: null,
   messages: null,
   chatId: null,
@@ -46,11 +50,49 @@ export const chatSlice = createSlice({
 		closeChatModal: state => {
 			state.modalProps.isOpenChat = false
 		},
-    
+  },
+    extraReducers: (builder) => {
+    // Add reducers for additional action types here, and handle loading state as needed
+    builder.addCase(fetchChats.fulfilled, (state, action) => {
+      console.log(action.payload);
+     return {...state, chats: action.payload}
+    })
   },
 });
 
 export const { setChats, setError, setMessages, setChatId,openChatModal,closeChatModal,addChatMember,removeChatMember } = chatSlice.actions;
+
+
+
+
+
+// export const fetchChats = (token) => async (dispatch) => {
+//   try {
+//     const chatsIDs = await getChats(userId, token);
+//     const chatPromises = chatsIDs.map(async (chatId) => {
+//       const data = await getChatMessages(chatId, token);
+//       let lastMessage = data.length > 0 ? data[data.length - 1] : { text: '', user: '' };
+
+//       if (data.length === 0) {
+//         const chatMembers = await getChatMembers(chatId, token);
+//         lastMessage.user = chatMembers.length > 1 ? chatMembers[1] : chatMembers[0];
+//       }
+
+//       return { id: chatId, lastMessage };
+//     });
+
+//     const updatedChats = await Promise.all(chatPromises);
+//     dispatch(setChats(updatedChats));
+//   } catch (error) {
+//     if (error.response) {
+//       dispatch(setError(`Error ${error.response.status}: ${error.response.data}`));
+//     } else if (error.request) {
+//       dispatch(setError('Error: no response'));
+//     } else {
+//       dispatch(setError(`Error: ${error.message}`));
+//     }
+//   }
+// };
 
 // export const fetchChats = (token) => async (dispatch) => {
 //   try {
@@ -76,135 +118,135 @@ export const { setChats, setError, setMessages, setChatId,openChatModal,closeCha
 //   }
 // };
 
-export const fetchChats = (token) => async (dispatch) => {
-  try {
-    const chatsIDs = await getChats(userId, token);
-    const chatPromises = chatsIDs.map(async (chatId) => {
-      const data = await getChatMessages(chatId, token);
-      let lastMessage = null
-      if(data.length > 0) {
-        lastMessage= data[data.length-1]
-      } else {
-        let chatMembers = await getChatMembers(chatId, token)
-         lastMessage = {
-          text: '',
-          user: chatMembers.length > 1 ?  chatMembers[1] : chatMembers[0]
-        }
-      }
-      return { id: chatId, lastMessage: lastMessage };
-    });
+// export const fetchChats = (token) => async (dispatch) => {
+//   try {
+//     const chatsIDs = await getChats(userId, token);
+//     const chatPromises = chatsIDs.map(async (chatId) => {
+//       const data = await getChatMessages(chatId, token);
+//       let lastMessage = null
+//       if(data.length > 0) {
+//         lastMessage= data[data.length-1]
+//       } else {
+//         let chatMembers = await getChatMembers(chatId, token)
+//          lastMessage = {
+//           text: '',
+//           user: chatMembers.length > 1 ?  chatMembers[1] : chatMembers[0]
+//         }
+//       }
+//       return { id: chatId, lastMessage: lastMessage };
+//     });
 
-    const updatedChats = await Promise.all(chatPromises);
-    dispatch(setChats(updatedChats));
+//     const updatedChats = await Promise.all(chatPromises);
+//     dispatch(setChats(updatedChats));
 
-  } catch (error) {
-    if (error.response) {
-      dispatch(
-        setError(`Error ${error.response?.status}: ${error.response?.data}`)
-      );
-    } else if (error.request) {
-      dispatch(setError("Error: no response"));
-    } else {
-      dispatch(setError(`Error: ${error?.message}`));
-    }
-  }
-};
-
-
-export async function getChatMembers(chatID, token) {
-  try {
-    const { data } = await axios.get(
-      `${
-        process.env.REACT_APP_SERVER_URL || ""
-      }/api/v1/get-members-chat/${chatID}`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
-
-export async function getChatMessages(chatID, token) {
-  try {
-    const { data } = await axios.get(
-      `${
-        process.env.REACT_APP_SERVER_URL || ""
-      }/api/v1/get-messages-chat/${chatID}`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
-
-export async function getChats(userId, token) {
-  try {
-    const { data } = await axios.get(
-      `${process.env.REACT_APP_SERVER_URL || ''}/api/v1/get-chats-user/${userId}`,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
+//   } catch (error) {
+//     if (error.response) {
+//       dispatch(
+//         setError(`Error ${error.response?.status}: ${error.response?.data}`)
+//       );
+//     } else if (error.request) {
+//       dispatch(setError("Error: no response"));
+//     } else {
+//       dispatch(setError(`Error: ${error?.message}`));
+//     }
+//   }
+// };
 
 
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
+// export async function getChatMembers(chatID, token) {
+//   try {
+//     const { data } = await axios.get(
+//       `${
+//         process.env.REACT_APP_SERVER_URL || ""
+//       }/api/v1/get-members-chat/${chatID}`,
+//       {
+//         headers: {
+//           Authorization: `Bearer ${token}`,
+//         },
+//       }
+//     );
 
-export async function createMessage(message, token) {
-  try {
+//     return data;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
+
+// export async function getChatMessages(chatID, token) {
+//   try {
+//     const { data } = await axios.get(
+//       `${
+//         process.env.REACT_APP_SERVER_URL || ""
+//       }/api/v1/get-messages-chat/${chatID}`,
+//       {
+//         headers: {
+//           Authorization: `Bearer ${token}`,
+//         },
+//       }
+//     );
+
+//     return data;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
+
+// export async function getChats(userId, token) {
+//   try {
+//     const { data } = await axios.get(
+//       `${process.env.REACT_APP_SERVER_URL || ''}/api/v1/get-chats-user/${userId}`,
+//       {
+//         headers: {
+//           Authorization: `Bearer ${token}`,
+//         },
+//       }
+//     );
+
+
+//     return data;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
+
+// export async function createMessage(message, token) {
+//   try {
     
-    const { data } = await axios.post(
-      `${process.env.REACT_APP_SERVER_URL || ""}/api/v1/message`,
+//     const { data } = await axios.post(
+//       `${process.env.REACT_APP_SERVER_URL || ""}/api/v1/message`,
 
-      message,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "application/json",
-        },
-      }
-    );
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
-export   async function createChat(chat, token) {
-  try {
+//       message,
+//       {
+//         headers: {
+//           Authorization: `Bearer ${token}`,
+//           "Content-Type": "application/json",
+//         },
+//       }
+//     );
+//     return data;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
+// export   async function createChat(chat, token) {
+//   try {
     
-    const { data } = await axios.post(
-      `${process.env.REACT_APP_SERVER_URL || ""}/api/v1/create-chat`,
+//     const { data } = await axios.post(
+//       `${process.env.REACT_APP_SERVER_URL || ""}/api/v1/create-chat`,
 
-      chat,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-          "Content-Type": "application/json",
-        },
-      }
-    );
-    return data;
-  } catch (error) {
-    throw error;
-  }
-}
+//       chat,
+//       {
+//         headers: {
+//           Authorization: `Bearer ${token}`,
+//           "Content-Type": "application/json",
+//         },
+//       }
+//     );
+//     return data;
+//   } catch (error) {
+//     throw error;
+//   }
+// }
 export  function getAllUsers() {
   try {
